bootstrap added, lotts of little changes just tweaks now tho thank goodness

+++ routes/web.php
@@ -16,26 +16,23 @@ Route::get('/', function () {
 });
 
 
-Route::get('posts/create', 'PostController@create')->name('posts.create');
+Route::get('posts/create', 'PostController@create')->name('posts.create')->middleware('auth');
 
-Route::post('posts/store', 'PostController@store')->name('post.store');
+Route::post('posts/store', 'PostController@store')->name('post.store')->middleware('auth');
  
-Route::post('posts/comment', 'CommentController@store')->name('comment.store');
+Route::post('posts/comment/{id}', 'CommentController@store')->name('comment.store')->middleware('auth');
 
 Route::get('posts/all', 'PostController@index')->name('posts.index');
 
 Route::get('posts/{id}', 'PostController@show')->name('posts.show');
 
-Route::delete('posts/{id}', 'PostController@destroy')->name('post.destroy');
+Route::delete('posts/{id}', 'PostController@destroy')->name('post.destroy')->middleware('auth');
 
 Route::get('users/{id}', 'UserController@show')->name('users.show');
 
-Route::post('users/{id}', 'UserController@friend')->name('users.friend');
+Route::post('users/{id}', 'UserController@friend')->name('users.friend')->middleware('auth');
 
-Route::get('users', 'UserController@myFriends')->name('users.friends');
-
-
-
+Route::get('users', 'UserController@myFriends')->name('users.friends')->middleware('auth');
 
 Auth::routes();
